+++ test/lib/GpioMock.mocks.js
@@ -1,0 +1,3 @@
+const GpiosMocks = require('./Gpios.mocks.js')
+
+module.exports = new GpiosMocks('./GpioMock')